fix: address review comments and extract more components
Review fixes:
- Fix React.ComponentType import by using LucideIcon type (StatChip, TodoItem)
- Add aria-expanded/aria-controls to CollapsibleSection for accessibility
- Fix wrong label key in Scan modal (labRefField → scanRefField)

Further file splitting:
- Extract StatChip component from DashboardScreen → src/components/StatChip.tsx
- Extract TodoItem component from DashboardScreen → src/components/TodoItem.tsx
- Extract FilmInfoSection → src/screens/film-detail/FilmInfoSection.tsx
- Extract FloatingActionBar → src/screens/film-detail/FloatingActionBar.tsx
- FilmDetailScreen.tsx now 339 lines (was 1587, then 477)
